Refactor postman_tasks.py and utils.py
